+++ docs/CHANGELOG.rst
@@ -20,6 +20,10 @@ Added
 Fixed
 -----
 - Enable all top-level class definitions in Python processes
+- Make filtering via foreign key more 'Django like': when foreign key does
+  not exist return empty set instead of raising validation exception.
+  Also when filtering using list of foreign keys do not raise validation
+  exception if some foreign keys in the list do not exist.
 
 
 ===================